% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_mat.R
\name{a_mat}
\alias{a_mat}
\title{Calculate age at maturity for all species}
\usage{
a_mat(params)
}
\arguments{
\item{params}{A MizerParams object or a species_params data frame}
}
\value{
A named vector. The names are the species names and the values are
the ages at maturity.
}
\description{
Uses the age at maturity that is implied by the von Bertalanffy growth
curve specified by the \code{w_inf}, \code{k_vb}, \code{t0} parameters in the
species_params data frame.
}
