% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundanceControl.R
\name{abundanceControl}
\alias{abundanceControl}
\alias{abundanceControlUI}
\title{Controlling egg abundance in the tuning gadget}
\usage{
abundanceControl(input, output, session, params, flags, ...)

abundanceControlUI(p, sp)
}
\arguments{
\item{input}{.}

\item{output}{.}

\item{session}{.}

\item{params}{.}

\item{flags}{.}

\item{...}{.}

\item{p}{The MizerParams object currently being tuned.}

\item{sp}{The single row from the pecies parameter data frame for the
currently selected species.}
}
\value{
A tagList with a slider for the egg density
}
\description{
Controlling egg abundance in the tuning gadget
}
