% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignResource.R
\name{alignResource}
\alias{alignResource}
\title{Rescale resource to be in line with fish community}
\usage{
alignResource(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
A MizerParams object with updated resource carrying capacity,
resource initial value, resource parameter lambda and species parameter
gamma.
}
\description{
This function sets the coefficient of the resource power law so that the
resource spectrum aligns approximately with the fish community spectrum and it
then rescales the search volumes of all fish species so that their resource
encounter rate stays the same. This will therefore modify the encounter rate
with other prey.
}
