% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass_callback.R
\name{biomass_callback}
\alias{biomass_callback}
\title{Callback function for plotting biomass in real-time}
\usage{
biomass_callback(sim, t_idx, ...)
}
\arguments{
\item{sim}{A \link[mizer:MizerSim]{MizerSim} object.}

\item{t_idx}{The current time index.}

\item{...}{Other arguments passed to the callback (such as \code{ylim}, \code{species},
\code{use_cutoff}, etc.).}
}
\description{
This function can be passed to the \code{callback} argument of \code{\link[mizer:project]{project()}} to
plot the biomasses of the species in real-time as the simulation runs.
}
\examples{
\dontrun{
params <- NS_params
# Open an external graphics window to see the real-time plot in RStudio
x11()
sim <- project(params, callback = biomass_callback)
}
}
