% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomassTab.R
\name{biomassTab}
\alias{biomassTab}
\alias{biomassTabUI}
\title{Biomass tab for tuning gadget}
\usage{
biomassTab(input, output, session, params, logs, trigger_update, ...)

biomassTabUI(params, ...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{trigger_update}{Reactive value used for triggering update of
species parameter sliders.}

\item{...}{Unused}
}
\description{
The Biomass tab shows:
\itemize{
\item A plot of total biomass for each species, compared to
observed biomasses when available, using \code{\link[=plotBiomassVsSpecies]{plotBiomassVsSpecies()}}.
\item Buttons "Calibrate" and "Match" that trigger a call to
\code{\link[=calibrateBiomass]{calibrateBiomass()}} or \code{\link[=matchBiomasses]{matchBiomasses()}} respectively.
}
}
\details{
Clicking on a species in the biomass plot makes that species the selected
species. Double-clicking on a species selects that species \strong{and}
changes its biomass.
}
