% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotYieldCurve.R
\name{calc_vals}
\alias{calc_vals}
\title{Calculates SSB, RDD and yield at a range of fishing mortalities}
\usage{
calc_vals(params, F_range, idx_species, tol = 0.001, t_max = 100)
}
\arguments{
\item{params}{An object of class \code{MizerParams}.}

\item{F_range}{A sequence of fishing mortalities at which to evaluate the
yield. If missing, it is set to
\code{seq(from = 0, to = F_max, length.out = no_steps)}.}

\item{tol}{The \code{projectToSteady} function stops when the relative change in the egg production
RDI over t_per years is less than tol for every species.}

\item{t_max}{The longest time to run project to find steady state.}
}
\value{
A data frame with columns \code{F}, \code{SY}, \code{SSB}, \code{RDD}
}
\description{
This function replaces a loop used multiple times within
\code{getYieldCurve}
}
