% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchTab.R
\name{catchTab}
\alias{catchTab}
\alias{catchTabUI}
\title{Serve tab with catch data}
\usage{
catchTab(
  input,
  output,
  session,
  params,
  logs,
  trigger_update,
  catch = NULL,
  ...
)

catchTabUI(...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{trigger_update}{Reactive value used for triggering update of
species parameter sliders.}

\item{catch}{Data frame holding binned observed catch data. The data can
be binned either into length bins or weight bins. In the former case the data
frame should have columns \code{length} and \code{dl} holding the start of
the size bins in cm and the width of the size bins in cm respectively. In
the latter case the data frame should have columns \code{weight} and
\code{dw} holding the start of the size bins in grams and the width of the
size bins in grams. The data frame also needs to have the columns
\code{species} (the name of the species), \code{catch} (the number of
individuals of a particular species caught in a size bin).}

\item{...}{Unused}
}
\description{
Serve tab with catch data
}
