% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparedietmatrix.R
\name{comparedietmatrix}
\alias{comparedietmatrix}
\title{I am not sure if this code works as intended. It appears so, as I have checked it,
but the results are not what it is expected.}
\usage{
comparedietmatrix(unharvestedprojection, harvestedprojection, timerange)
}
\arguments{
\item{harvested}{A mizerSim object}

\item{unharvested}{A mizerSim object - to compare to.}

\item{chosentime}{The year range to plot (example 1:2).}
}
\value{
A ggplot object of a matrix of predator species on the X axis,
prey species on the Y axis. The colour of the box indicates the change
of the proportion in the predator's diet of the given prey species.
}
\description{
I will also need to see if edge cases, where the timerange is 1:1, work.
Plot Relative Diet Proportion of each Prey/Predator
}
\details{
This function takes two mizerSim objects and calculates the relative
change in the proportion of a given prey species in a predators diet. This
is done for every prey/predator in the model.
}
\examples{
harvested <- getBiomass(NS_sim)
unharvested <- getBiomass(NS_sim)
comparedietmatrix(harvested, unharvested, 5)

}
