% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareParams.R
\name{compareParams}
\alias{compareParams}
\title{Compare two MizerParams objects and print out differences}
\usage{
compareParams(params1, params2)
}
\arguments{
\item{params1}{First MizerParams object}

\item{params2}{Second MizerParams object}
}
\description{
Compare two MizerParams objects and print out differences
}
\examples{
sp1 <- NS_species_params
params1 <- newMultispeciesParams(sp1)
sp2 <- sp1
sp2$w_mat[1] <- 10
params2 <- newMultispeciesParams(sp2)
compareParams(params1, params2)
}
