% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietTab.R
\name{dietTab}
\alias{dietTab}
\alias{dietTabUI}
\title{Serve tab with diet plot}
\usage{
dietTab(input, output, session, params, logs, diet = NULL, ...)

dietTabUI(...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{...}{Unused}
}
\description{
Serve tab with diet plot
}
