% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectToSteady.R
\name{distanceLogN}
\alias{distanceLogN}
\title{Measure mean squared distance between log(N) in current and previous state}
\usage{
distanceLogN(params, current, previous)
}
\arguments{
\item{current}{A named list with entries \code{n}, \code{n_pp} and \code{n_other}
describing the current state}

\item{previous}{A named list with entries \code{n}, \code{n_pp} and \code{n_other}
describing the previous state}
}
\value{
A number expressing a distance between current and previous state.
}
\description{
This function can be used in \code{\link[=projectToSteady]{projectToSteady()}} to decide when sufficient
convergence to steady state has been achieved.
}
\seealso{
Other distance functions: 
\code{\link{distanceRDI}()}
}
\concept{distance functions}
