% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectToSteady.R
\name{distanceRDI}
\alias{distanceRDI}
\title{Measure distance between current and previous state in terms of RDI}
\usage{
distanceRDI(current, previous)
}
\arguments{
\item{current}{A named list with entries \code{n}, \code{n_pp}, \code{n_other} and
\code{rates} describing the current state}

\item{previous}{A named list with entries \code{n}, \code{n_pp}, \code{n_other} and
\code{rates} describing the previous state}
}
\value{
A number expressing a distance between current and previous state.
}
\description{
Measure distance between current and previous state in terms of RDI
}
