% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramsCompar.R
\name{getDiffParams}
\alias{getDiffParams}
\title{Function to generate the html with paramsCompar output}
\usage{
getDiffParams(folder)
}
\arguments{
\item{folder}{Character string pointing to the folder
where the params object to compare are stored}
}
\description{
Function to generate the html with paramsCompar output
}
