% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYieldVsFcurve.R
\name{getYieldVsFcurve}
\alias{getYieldVsFcurve}
\title{Calculate a yield-vs-F curve for a species.}
\usage{
getYieldVsFcurve(params, ixSpecies, nSteps = 10, Fmax = 3, bPlot = FALSE)
}
\arguments{
\item{params}{An object of class \code{MizerParams}.}

\item{ixSpecies}{The species number to make the calculation over}

\item{nSteps}{The number of steps in F to calculate}

\item{Fmax}{The maximum fishing mortality}

\item{bPlot}{Boolean that indicates whether a plot is to be made}

\item{Frange}{The range of fishing mortalities to run over}
}
\value{
An list with yields and fishing mortalities
}
\description{
Calculates the yield of a species for a range of fishing mortalities
for species across all gears at each
simulation time step. \strong{Note:} The function just returns the yield
at the last time step of the simulation, which might not have converged,
or might oscillate.
}
\examples{
\dontrun{
params <- newMultispeciesParams(NS_species_params_gears, inter)
y <- getYieldVsFcurve(params,11,bPlot=TRUE)
}
}
\concept{summary functions}
\concept{summary_function}
