% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthControl.R
\name{growthControl}
\alias{growthControl}
\alias{growthControlUI}
\title{A single slider that adjusts both \code{h} and \code{gamma}}
\usage{
growthControl(input, output, session, params, flags, ...)

growthControlUI(p, sp)
}
\arguments{
\item{input}{.}

\item{output}{.}

\item{session}{.}

\item{params}{.}

\item{flags}{.}

\item{...}{.}

\item{p}{The MizerParams object currently being tuned.}

\item{sp}{The single row from the pecies parameter data frame for the
currently selected species.}
}
\description{
A single slider that adjusts both \code{h} and \code{gamma}
}
