% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guildplot.R
\name{guildplot}
\alias{guildplot}
\title{Plot Guild Relative Change Across Timescales}
\usage{
guildplot(harvestedprojection, unharvestedprojection, year1, year2)
}
\arguments{
\item{year1}{The lower year to plot in the range}

\item{year2}{The higher year to plot in the range}

\item{harvested}{A mizerSim object}

\item{unharvested}{A mizerSim object - to compare to.}
}
\value{
A ggplot object that plots 3 bars per species - in the short,
chosen and long time - it plots the relative biomass of each feeding guild
in comparison to the unharvested.
}
\description{
This function takes two mizerSim objects and calculates the relative \%
change in each given feeding guilde in the chosen year, short term (1/2 of the
chosen year) and the long term (2x the chosen year)
This function requires a dataframe in the environment titled guildparams - this dataframe should have
a column for minw (minimum weight of guild), maxw (maxmimum weight), guild (
the guild for the given weight), and a column for the species (which matches the mizersim species).
The mizerSim objects must also have a tmax of 2 * year2.
}
\examples{
harvested <- getBiomass(NS_sim)
unharvested <- getBiomass(NS_sim)
guildplot(harvested, unharvested, 1, 2)

}
