% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homeTab.R
\name{homeTab}
\alias{homeTab}
\alias{homeTabUI}
\title{tuneParams home tab}
\usage{
homeTab(input, output, session, params, logs, trigger_update, ...)

homeTabUI(params, help = TRUE, ...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{trigger_update}{Reactive value used for triggering update of
species parameter sliders.}

\item{...}{Unused}

\item{help}{Boolean. If FALSE then the help text is not included on the tab.
This is useful when including this tab as an element of another tab.}
}
\description{
This tab contains a brief introduction to the app
and the sponsors
}
