% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulateParams.R
\name{markBackground}
\alias{markBackground}
\title{Designate species as background species}
\usage{
markBackground(object, species = NULL)
}
\arguments{
\item{object}{An object of class \linkS4class{MizerParams} or
\linkS4class{MizerSim}.}

\item{species}{The species to be selected. Optional. By default all target
species are selected. A vector of species names, or a
numeric vector with the species indices, or a logical vector indicating for
each species whether it is to be selected (TRUE) or not.}
}
\value{
An object of the same class as the \code{object} argument
}
\description{
Marks the specified set of species as background species. Background species
are handled differently in some plots and their abundance is automatically
adjusted in \code{\link[=adjustBackgroundSpecies]{adjustBackgroundSpecies()}} to keep the community close to the
Sheldon spectrum.
}
\examples{
\dontrun{
params <- newMultispeciesParams(NS_species_params_gears, inter)
sim <- project(params, effort=1, t_max=20, t_save = 0.2, progress_bar = FALSE)
sim <- markBackground(sim, species = c("Sprat", "Sandeel",
                                       "N.pout", "Dab", "Saithe"))
plotSpectra(sim)
}
}
