% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYieldVsFcurve.R
\name{maxFthreshold}
\alias{maxFthreshold}
\title{Determines maximum the threshold where
fisheries effort should stop}
\usage{
maxFthreshold(params, idx_species, threshold_var)
}
\arguments{
\item{params}{An object of class \code{MizerParams}.}

\item{idx_species}{TODO: document}

\item{threshold_var}{A numeric value used in max_func to calculate the threshold
from the \code{params} object. If supplied the function returns the value of the
threshold using \code{threshold_var}. If not supplied the function returns the
unchanged value, which can be compared against the version which uses \code{threshold_var}.}
}
\value{
a numeric value
}
\description{
This function creates a threshold signaling \code{getMaxF} when
effort is considered at it's maximum. In this case it is
when the fished biomass is less than 10\% of the initial
biomass exposed to fisheries.
}
