% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizerExperimental-package.R
\docType{package}
\name{mizerExperimental-package}
\alias{mizerExperimental}
\alias{mizerExperimental-package}
\title{mizerExperimental: Extends the mizer package with experimental features}
\description{
This is an extension package for the
\href{https://sizespectrum.org/mizer/}{mizer package}
that contains additional features
contributed by the mizer community. These features can then be further
improved while being used by the community. Once matured, frequently-used
features can be moved to the core mizer package.

If you have code for working with mizer that you feel might be useful to
other mizer users, please describe it in
\href{https://github.com/sizespectrum/mizerExperimental/issues/new}{a new issue on the issue tracker}
and we'll comment on how it can be included in this package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sizespectrum/mizerExperimental}
  \item Report bugs at \url{https://github.com/sizespectrum/mizerExperimental/issues}
}

}
\keyword{internal}
