% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{mizerExperimental}
\alias{mizerExperimental}
\alias{mizerExperimental-package}
\title{mizerExperimental: Extends the mizer package with experimental features}
\description{
mizerExperimental: Extends the mizer package with experimental features
}
