% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberTab.R
\name{numberTab}
\alias{numberTab}
\alias{numberTabUI}
\title{Number tab for tuning gadget}
\usage{
numberTab(
  input,
  output,
  session,
  params,
  params_old,
  logs,
  trigger_update,
  ...
)

numberTabUI(params, ...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{params_old}{Reactive value holding non-updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{trigger_update}{Reactive value used for triggering update of
species parameter sliders.}

\item{...}{Unused}
}
\description{
The Number tab shows:
\itemize{
\item A plot of total number for each species, compared to
observed numbers when available, using \code{\link[=plotNumberVsSpecies]{plotNumberVsSpecies()}}.
\item Buttons "Calibrate" and "Match" that trigger a call to
\code{\link[=calibrateNumber]{calibrateNumber()}} or \code{\link[=matchNumbers]{matchNumbers()}} respectively.
}
}
\details{
Clicking on a species in the number plot makes that species the selected
species. Double-clicking on a species selects that species \strong{and}
changes its number.
}
