% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramsCompar.R
\name{paramsCompar}
\alias{paramsCompar}
\title{function allowing the comparison of multiple Mizer params object downloaded from tuneParams
need to compare species param and resource param on a plot}
\usage{
paramsCompar(folder)
}
\arguments{
\item{folder}{folder name where all the zip files are}
}
\description{
instructions
markdown need to look like:
table with different parameters
followed by biomass and size spectrum plot
}
