% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBiomassVsSpecies.R
\name{plotBiomassVsSpecies}
\alias{plotBiomassVsSpecies}
\title{Plot the biomass against species}
\usage{
plotBiomassVsSpecies(params, ...)
}
\arguments{
\item{params}{A MizerParams object}

\item{...}{Not used.}
}
\description{
This is a generic function with a method for objects of class
\link[mizer:MizerParams]{MizerParams}.
}
\examples{
plotBiomassVsSpecies(NS_params)

}
