% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discards.R
\name{plotCatchVsSize}
\alias{plotCatchVsSize}
\alias{plotlyCatchVsSize}
\title{Plot the size distribution of the catch}
\usage{
plotCatchVsSize(
  object,
  species = NULL,
  gears = NULL,
  catch = NULL,
  x_var = c("Weight", "Length"),
  return_data = FALSE
)

plotlyCatchVsSize(
  object,
  species = NULL,
  gear = NULL,
  catch = NULL,
  x_var = c("Weight", "Length"),
  ...
)
}
\arguments{
\item{object}{An object of class \linkS4class{MizerSim} or
\linkS4class{MizerParams}.}

\item{species}{The name of the species for which to plot the catches.}

\item{gears}{Optional. The name of a gear or a list of names of gear. If
supplied, only the catches from these gears will be included. Otherwise
catches from all gears will be included.}

\item{catch}{Data frame holding binned observed catch data. See Details,}

\item{x_var}{Determines whether to show the size distribution of the catch as
a function of weight ("Weight") or as a function of length ("Length").
Default is "Weight".}
}
\value{
A ggplot2 object, unless \code{return_data = TRUE}, in which case a list
composed of two slots is returned. First slot is a data frame with the four
variables 'w' or 'l' (depending on \code{x_var}), 'Catch density', 'Type', 'Species
and the second slot is a data frame with the four variables 'w_mat',
'Species', 'y_coord', 'Type' (to plot vertical lines).
}
\description{
Plots the normalised number density of the catch for a species as a function
of either length or weight. In addition to the catch in the model, the
observed catch will be plotted if it is supplied via the \code{catch} argument.
Also superimposes a plot of the number density of all individuals of the
species.
}
\details{
The observed catch is provided via the \code{catch} data frame. The catch data can
be binned either into length bins or weight bins. In the former case the data
frame should have columns \code{length} and \code{dl} holding the start of
the size bins in cm and the width of the size bins in cm respectively. In the
latter case the data frame should have columns \code{weight} and \code{dw}
holding the start of the size bins in grams and the width of the size bins in
grams.

The \code{catch} data frame also needs to have the columns \code{species} (the
name of the species), \code{gear} (the name of the gear) and \code{catch}
(the number of individuals of a particular species caught by a particular
gear in a size bin).

Optionally, the \code{catch} data frame can have an additional column called
\code{discarded}, which hold the number of individuals of a
particular species that were discarded by a particular gear in a
size bin. If such discard data is available, then the function will plot
three curves representing the observed data: "Observed catch", "Observed
discards" and "Observed landings", where the landings and the discards add
up to the observed catch.
}
\examples{
\donttest{
plotCatchVsSize(NS_params, species = "Cod")

# Returning the data frame
fr <- plotCatchVsSize(NS_params, species = "Cod", return_data = TRUE)
str(fr)
}
}
\seealso{
Other plotting functions: 
\code{\link{plotDeath}()},
\code{\link{plotEnergyBudget}()},
\code{\link{plotResourceLevel}()},
\code{\link{plotResourcePred}()},
\code{\link{plotYieldVsF}()},
\code{\link{plotYieldVsSize}()}
}
\concept{plotting functions}
