% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDeath}
\alias{plotDeath}
\alias{plotlyDeath}
\title{Plot the sources of external, predation and fishing mortality
per species and size}
\usage{
plotDeath(
  object,
  species = NULL,
  proportion = TRUE,
  return_data = FALSE,
  xtrans = c("log10", "identity"),
  ...
)

plotlyDeath(object, species = NULL, proportion = TRUE, ...)
}
\arguments{
\item{object}{An object of class \linkS4class{MizerSim} or
\linkS4class{MizerParams}.}

\item{species}{The name of the predator species for which to plot the
mortality.}

\item{proportion}{A boolean value that determines whether values should be
displayed as proportions from 0 to 1 or with their actual values. Default
is TRUE.}

\item{return_data}{A boolean value that determines whether the formatted data
used for the plot is returned instead of the plot itself. Default value is
FALSE}

\item{xtrans}{The transformation to apply to the x-axis. Either "log10"
(default) or "identity".}

\item{...}{Other arguments (currently unused)}
}
\value{
A ggplot2 object, unless \code{return_data = TRUE}, in which case a data
frame with the four variables 'w', 'value', 'Cause', 'Species' is returned.
}
\description{
This is a generic function with methods for objects of class
\linkS4class{MizerSim} and \linkS4class{MizerParams}. The method for a
\code{MizerSim} object plots the mortality in the initial state of the
simulation. An interactive version is available as \code{plotlyDeath()}.
}
\examples{
\donttest{
plotDeath(NS_params, species = "Cod")

# Returning the data frame
fr <- plotDeath(NS_params, species = "Cod", return_data = TRUE)
str(fr)
}
}
\seealso{
\link[mizer:plotting_functions]{plotting_functions}

Other plotting functions:
\code{\link[=plotEnergyBudget]{plotEnergyBudget()}},
\code{\link[=plotResourceLevel]{plotResourceLevel()}},
\code{\link[=plotResourcePred]{plotResourcePred()}},
\code{\link[=plotYieldVsSize]{plotYieldVsSize()}}
}
\concept{plotting functions}
