% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotEnergyBudget}
\alias{plotEnergyBudget}
\alias{plotlyEnergyBudget}
\title{Plot the energy budget of each species through size.}
\usage{
plotEnergyBudget(
  object,
  species = NULL,
  logarithmic = TRUE,
  return_data = FALSE,
  ...
)

plotlyEnergyBudget(object, species = NULL, logarithmic = TRUE, ...)
}
\arguments{
\item{object}{An object of class \linkS4class{MizerSim} or
\linkS4class{MizerParams}.}

\item{species}{The name of the predator species for which to plot the
mortality.}

\item{logarithmic}{A boolean value that determines whether values should be
displayed logarithmicly or linearly. Default is TRUE.}

\item{return_data}{A boolean value that determines whether the formatted data
used for the plot is returned instead of the plot itself. Default value is
FALSE}

\item{...}{Other arguments (currently unused)}
}
\value{
A ggplot2 object, unless \code{return_data = TRUE}, in which case a list composed of
two slots is returned. First slot is a data frame with the four variables 'w', 'value',
'Type', 'Species and the second slot is a data frame with the five variables 'w_mat',
'w_max', 'Species', 'y_coord', 'Type' (to plot vertical lines).
}
\description{
This budget is divided between growth, income, metabolic loss and reproduction.
}
\details{
This is a generic function with methods for objects of class
\linkS4class{MizerSim} and \linkS4class{MizerParams}. The method for a
\code{MizerSim} object plots the energy budget in the initial state of the
simulation. An interactive version is available as \code{plotlyEnergyBudget()}.
}
\examples{
\donttest{
plotEnergyBudget(NS_params, species = "Cod")

# Returning the data frame
fr <- plotEnergyBudget(NS_params, return_data = TRUE)
str(fr)
}
}
\seealso{
\link[mizer:plotting_functions]{plotting_functions}

Other plotting functions:
\code{\link[=plotDeath]{plotDeath()}},
\code{\link[=plotResourceLevel]{plotResourceLevel()}},
\code{\link[=plotResourcePred]{plotResourcePred()}},
\code{\link[=plotYieldVsSize]{plotYieldVsSize()}}
}
\concept{plotting functions}
