% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotYieldCurve.R
\name{plotYieldCurve}
\alias{plotYieldCurve}
\title{Plot normalised SBB, RDD and sustainable yield versus F}
\usage{
plotYieldCurve(
  params,
  species,
  no_steps = 20,
  F_max,
  F_range,
  tol = 0.001,
  t_max = 100
)
}
\arguments{
\item{params}{An object of class \code{MizerParams}.}

\item{species}{Name of the target species}

\item{no_steps}{The number of steps to use. Only used if \code{F_range} is
missing.}

\item{F_max}{The maximum fishing mortality. Used only if \code{F_range} is
missing.}

\item{F_range}{A sequence of fishing mortalities at which to evaluate the
SSB. If missing, it is set to
\code{seq(from = 0, to = F_max, length.out = no_steps)}.}

\item{tol}{The \code{projectToSteady} function stops when the relative change in
the egg production RDI over t_per years is less than tol for every species.}

\item{t_max}{The longest time to run project to find steady state.}
}
\value{
A ggplot object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculates SSB, RDI, RDD, and the yield of a species for a range of
fishing mortalities for that species while the fishing mortalities for the
other species are held fixed.
}
\examples{
\dontrun{
params <- newMultispeciesParams(NS_species_params_gears, inter)
plotYieldCurve(params, "Cod")
}
}
\seealso{
getYieldVsF

Other plotting functions: 
\code{\link{plotDeath}()},
\code{\link{plotEnergyBudget}()},
\code{\link{plotResourceLevel}()},
\code{\link{plotResourcePred}()},
\code{\link{plotYieldVsF}()},
\code{\link{plotYieldVsSize}()}
}
\concept{plotting functions}
