% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotYieldVsSpecies.R
\name{plotYieldVsSpecies}
\alias{plotYieldVsSpecies}
\title{Plot the yield against species}
\usage{
plotYieldVsSpecies(params, gear = NULL, ...)
}
\arguments{
\item{params}{A MizerParams object}

\item{gear}{Optional. The name of a gear. If supplied, only the yield from
this gear will be displayed.}

\item{...}{Not used.}
}
\description{
This is a generic function with a method for objects of class
\link[mizer:MizerParams]{MizerParams}.
}
\examples{
plotYieldVsSpecies(NS_params)

}
