% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predationControl.R
\name{predationControl}
\alias{predationControl}
\alias{predationControlUI}
\title{Controlling the predation parameters in the tuning gadget}
\usage{
predationControl(input, output, session, params, flags, ...)

predationControlUI(p, sp)
}
\arguments{
\item{input}{.}

\item{output}{.}

\item{session}{.}

\item{params}{.}

\item{flags}{.}

\item{...}{.}

\item{p}{The MizerParams object currently being tuned.}

\item{sp}{The single row from the pecies parameter data frame for the
currently selected species.}
}
\description{
Controlling the predation parameters in the tuning gadget
}
