% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratesTab.R
\name{ratesTab}
\alias{ratesTab}
\alias{ratesTabUI}
\title{Serve tab with rates plots}
\usage{
ratesTab(input, output, session, params, logs, ...)

ratesTabUI(...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding a MizerParams object.}

\item{logs}{Environment holding the log of steady states.}

\item{...}{Unused}
}
\description{
Serve tab with rates plots
}
