% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproTab.R
\name{reproTab}
\alias{reproTab}
\alias{reproTabUI}
\title{Serve tab with plot of reproductive success rates}
\usage{
reproTab(input, output, session, params, logs, ...)

reproTabUI(...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{...}{Unused}
}
\description{
Serve tab with plot of reproductive success rates
}
