% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulateParams.R
\name{rescaleAbundance}
\alias{rescaleAbundance}
\title{Rescale Abundance}
\usage{
rescaleAbundance(params, factor)
}
\arguments{
\item{params}{A mizer params object}

\item{factor}{The factor by which the abundance of each species is multiplied.
This can be specified in two ways:
\itemize{
\item A named numeric vector where the name indicates the species and the
value gives the factor for that species. Only the named species are
affected.
\item  A number that gives the factor for all foreground species.
}}
}
\value{
An object of type \linkS4class{MizerParams}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Multiplies the abundances of all or of selected species by given factors and
then retunes the reproductive efficiencies accordingly.

Does not run the system to steady state. For that you should call
\code{\link[=steady]{steady()}} explicitly afterwards.
}
