% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceControl.R
\name{resourceControl}
\alias{resourceControl}
\alias{resourceControlUI}
\title{Controlling the resource parameters in the tuning gadget}
\usage{
resourceControl(input, output, session, params, flags, ...)

resourceControlUI(p, input)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{flags}{Environment holding flags to skip certain observers}

\item{...}{Unused}

\item{p}{The MizerParams object currently being tuned.}
}
\description{
Controlling the resource parameters in the tuning gadget
}
