% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceTab.R
\name{resourceTab}
\alias{resourceTab}
\alias{resourceTabUI}
\title{Serve tab with resource plots}
\usage{
resourceTab(input, output, session, params, logs, ...)

resourceTabUI(...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding a MizerParams object.}

\item{logs}{Environment holding the log of steady states.}

\item{...}{Unused}
}
\description{
Serve tab with resource plots
}
