% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SEAparams}
\alias{SEAparams}
\title{params object for the Tasman sea}
\format{
a MizerParams class object
}
\usage{
SEAparams
}
\description{
params object for the Tasman sea
}
\keyword{datasets}
