% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setResourceSemichemostat.R
\name{setResourceSemichemostat}
\alias{setResourceSemichemostat}
\title{Set semichemostat dynamics for resources}
\usage{
setResourceSemichemostat(params, resource_params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
A MizerParams object with updated resources
}
\description{
Uses the resource parameters provided in \code{resource_params} to set the
initial resource abundances and the resource rate and then sets the
resource capacity so that the initial abundances are at steady state, given
the current consumption rate.
}
