% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectraTab.R
\name{spectraTab}
\alias{spectraTab}
\alias{spectraTabUI}
\title{tuneParams tab showing size spectra}
\usage{
spectraTab(input, output, session, params, logs, trigger_update, ...)

spectraTabUI(params, help = TRUE, ...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{trigger_update}{Reactive value used for triggering update of
species parameter sliders.}

\item{...}{Unused}

\item{help}{Boolean. If FALSE then the help text is not included on the tab.
This is useful when including this tab as an element of another tab.}
}
\description{
This tab shows:
}
\details{
\itemize{
\item a plot of the biomass density as a function of log size,
including all species and the resource.
\item A slider for scaling the background (resource and possibly background
species).
\item If there are background species, buttons to adjust or remove those
background species.
}
}
