% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stomachTab.R
\name{stomachTab}
\alias{stomachTab}
\title{Serve tab with stomach data}
\usage{
stomachTab(input, output, session, params, logs, stomach, ...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{logs}{Environment holding the log of steady states.}

\item{stomach}{Data frame holding observations of prey items in predator
stomachs. The required columns are
\itemize{
\item \code{species} holding the name of the predator species,
\item \code{wpredator} with the weight in grams of the predator,
\item \code{wprey} with the weight of the prey item.
}
In case prey items of the same weight have been aggregated in the data
frame then there should be a column \code{Nprey} saying how many prey
items have been aggregated in each row.}

\item{...}{Unused}
}
\description{
Serve tab with stomach data
}
