% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneParams.R
\name{tuneGrowth}
\alias{tuneGrowth}
\title{Tune Growth}
\usage{
tuneGrowth(...)
}
\arguments{
\item{...}{All arguments are forwarded to \code{mizerEcopath::tuneGrowth()}.}
}
\value{
The tuned MizerParams object
}
\description{
This function has moved to the mizerEcopath package. If mizerEcopath is
installed, the call is forwarded automatically. Otherwise you will be
prompted to install mizerEcopath.
}
