% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneParams.R
\name{tuningGadget}
\alias{tuningGadget}
\title{Launch a customisable shiny gadget for tuning mizer parameters}
\usage{
tuningGadget(...)
}
\arguments{
\item{...}{All arguments are forwarded to \code{mizerEcopath::tuningGadget()}.}
}
\value{
The tuned MizerParams object
}
\description{
This function has moved to the mizerEcopath package. If mizerEcopath is
installed, the call is forwarded automatically. Otherwise you will be
prompted to install mizerEcopath.
}
\keyword{internal}
