% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulateParams.R
\name{updateInitialValues}
\alias{updateInitialValues}
\alias{updateInitialValues.MizerParams}
\title{Update the initial values}
\usage{
updateInitialValues(params, ...)

\method{updateInitialValues}{MizerParams}(params, ...)
}
\arguments{
\item{params}{A MizerParams object}

\item{...}{Not used.}
}
\value{
The MizerParams object with updated \code{initial_n} and
\code{initial_n_pp} slots.
}
\description{
Recalculates the steady-state abundances in a fixed background
given by the current abundances, keeping the abundances fixed in the
smallest size class for each species. Then readjusts the \code{erepro}
values.
}
