% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validSim.R
\name{validSim}
\alias{validSim}
\title{Returns a valid simulation by removing non-finite results}
\usage{
validSim(sim)
}
\arguments{
\item{sim}{A MizerSim object}
}
\value{
A MizerSim object with possibly fewer time steps
}
\description{
Checks whether any abundances are non-finite and if any are found, a warning
is issued and the simulation is truncated at the last time step where all
results are finite.
}
