% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundanceControl.R
\name{abundanceControl}
\alias{abundanceControl}
\alias{abundanceControlUI}
\title{Controlling egg abundance in the tuning gadget}
\usage{
abundanceControl(input, output, session, params, params_old, flags, ...)

abundanceControlUI(p, sp)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{params_old}{Reactive value holding non-updated MizerParams object}

\item{flags}{Environment holding flags to skip certain observers}

\item{...}{Unused}

\item{p}{The MizerParams object currently being tuned.}

\item{sp}{The single row from the pecies parameter data frame for the
currently selected species.}
}
\value{
A tagList with a slider for the egg density
}
\description{
Controlling egg abundance in the tuning gadget
}
