% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundanceTab.R
\name{abundanceTab}
\alias{abundanceTab}
\alias{abundanceTabUI}
\title{Abundance tab for tuning gadget}
\usage{
abundanceTab(input, ...)

abundanceTabUI(params, ...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{...}{Unused}

\item{params}{Reactive value holding updated MizerParams object}
}
\description{
The Abundance tab combines the \code{\link[=biomassTab]{biomassTab()}} and the \code{\link[=numberTab]{numberTab()}}
on a single tab.
}
