% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_mat.R
\name{age_mat_vB}
\alias{age_mat_vB}
\title{Calculate age at maturity from von Bertalanffy growth parameters}
\usage{
age_mat_vB(params)
}
\arguments{
\item{params}{A MizerParams object or a species_params data frame}
}
\value{
A named vector. The names are the species names and the values are
the ages at maturity.
}
\description{
Uses the age at maturity that is implied by the von Bertalanffy growth curve
specified by the \code{w_inf}, \code{k_vb}, \code{t0}, \code{a} and \code{b} parameters in the
species_params data frame.
}
\details{
If any of \code{a}, \code{b} or \code{k_vb} is missing for a species, the function returns
NA for that species. If \code{t0} is missing a default of 0 is used.
}
