% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animateSpectra.R
\name{animateSpectra}
\alias{animateSpectra}
\title{Animation of the abundance spectra}
\usage{
animateSpectra(
  sim,
  species,
  wlim = c(NA, NA),
  ylim = c(NA, NA),
  power = 1,
  total = FALSE,
  plankton = TRUE
)
}
\arguments{
\item{sim}{A MizerSim object}
}
\description{
Animation of the abundance spectra
}
\examples{
sim <- project(NS_params, t_max = 15, effort = 1)

}
\concept{plotting functions}
