% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneParams_biomassTab.R
\name{biomassTabUI}
\alias{biomassTabUI}
\alias{biomassTab}
\title{Biomass tab for tuning gadget}
\usage{
biomassTabUI(params, ...)

biomassTab(input, output, session, params, logs, trigger_update, ...)
}
\arguments{
\item{params}{.}

\item{...}{.}

\item{input}{.}

\item{output}{.}

\item{session}{.}

\item{logs}{.}

\item{trigger_update}{.}
}
\description{
The Biomass tab shows:
\itemize{
\item A plot of total biomass for each species, compared to
observed biomasses when available, using \code{\link[=plotBiomossObservedVsModel]{plotBiomossObservedVsModel()}}.
\item Buttons "Calibrate" and "Match" that trigger a call to
\code{\link[=calibrateBiomass]{calibrateBiomass()}} or \code{\link[=matchBiomasses]{matchBiomasses()}} respectively.
\item A plot of the biomass density spectra as a function of log size,
using \code{\link[=plotSpectra]{plotSpectra()}}.
}
}
