% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYieldVsFcurve.R
\name{distanceSSLogYield}
\alias{distanceSSLogYield}
\title{Measure distance between current and previous state in terms of yield.}
\usage{
distanceSSLogYield(params, current, previous, criterion = "SSE")
}
\arguments{
\item{params}{MizerParams}

\item{current}{A named list with entries \code{n}, \code{n_pp} and \code{n_other}
describing the current state}

\item{previous}{A named list with entries \code{n}, \code{n_pp} and \code{n_other}
describing the previous state}

\item{criterion}{TODO: document}
}
\value{
proportional difference between current and previous state
}
\description{
Calculates the proportional difference between getYield() outputs of current
and previous state. This function can be used in projectToSteady() to decide
when sufficient convergence to steady state has been achieved.
}
\concept{distance functions}
