% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishingControl.R
\name{fishingControl}
\alias{fishingControl}
\alias{fishingControlUI}
\title{Controlling the fishing parameters in the tuning gadget}
\usage{
fishingControl(input, output, session, params, flags, ...)

fishingControlUI(p, sp)
}
\arguments{
\item{input}{.}

\item{output}{.}

\item{session}{.}

\item{params}{.}

\item{flags}{.}

\item{...}{.}

\item{p}{The MizerParams object currently being tuned.}

\item{sp}{The single row from the pecies parameter data frame for the
currently selected species.}
}
\description{
Controlling the fishing parameters in the tuning gadget
}
