% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishingControl.R
\name{fishingControl}
\alias{fishingControl}
\alias{fishingControlUI}
\title{Controlling the fishing parameters in the tuning gadget}
\usage{
fishingControl(input, output, session, params, params_old, flags, ...)

fishingControlUI(p, sp)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{params_old}{Reactive value holding non-updated MizerParams object}

\item{flags}{Environment holding flags to skip certain observers}

\item{...}{Unused}

\item{p}{The MizerParams object currently being tuned.}

\item{sp}{The single row from the pecies parameter data frame for the
currently selected species.}
}
\description{
Controlling the fishing parameters in the tuning gadget
}
