% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYieldVsFcurve.R
\name{getMaxF}
\alias{getMaxF}
\title{Maximum viable fisheries effort for species}
\usage{
getMaxF(params, idx_species, effort_it = 1)
}
\arguments{
\item{params}{MizerParams}

\item{idx_species}{species row number in \code{params@species_params}}

\item{effort_it}{The width of steps used to find maximum fisheries mortality,
higher values means a faster calculation but less accurate maximum.}
}
\value{
dataframe
}
\description{
Calculate the maximum fisheries effort before reaching a specific threshold
Default case is biomass exposed to fisheries lower than 10\% of original biomass
}
