% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYieldVsFcurve.R
\name{getYieldVsF}
\alias{getYieldVsF}
\title{Calculate a yield-vs-F curve}
\usage{
getYieldVsF(params, species, F_range, no_steps = 10, effort_it = 1)
}
\arguments{
\item{params}{An object of class \code{MizerParams}.}

\item{species}{Name of the target species}

\item{F_range}{A sequence of fishing mortalities at which to evaluate the
yield. If missing, it is set to
\code{seq(from = 0, to = F_max, length.out = no_steps)}.}

\item{no_steps}{The number of steps to use. Only used if \code{F_range} is missing.}

\item{effort_it}{The width of steps used to find maximum fisheries mortality,
higher values means a faster calculation but less accurate maximum.}

\item{F_max}{The maximum fishing mortality. Used only if \code{F_range} is missing.}
}
\value{
A data frame with columns \code{F} and \code{yield}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculates the yield of a species for a range of fishing mortalities for
that species while the fishing mortalities for the other species are held
fixed.
}
\examples{
\dontrun{
params <- newMultispeciesParams(NS_species_params_gears, inter)
y <- getYieldVsF(params, "Cod", F_max = 1, no_steps = 5)
}
}
\seealso{
plotYieldVsF
}
\concept{summary functions}
\concept{summary_function}
