% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYieldVsFcurve.R
\name{getYieldVsF}
\alias{getYieldVsF}
\title{Calculate a yield-vs-F curve}
\usage{
getYieldVsF(
  params,
  species,
  F_range,
  F_max = 1,
  F_min = 0,
  no_steps = 10,
  distance_func = distanceSSLogN,
  tol = 0.001,
  t_max = 100
)
}
\arguments{
\item{params}{An object of class \code{MizerParams}.}

\item{species}{Name of the target species}

\item{F_range}{A sequence of fishing mortalities at which to evaluate the
yield. If missing, it is set to
\code{seq(from = 0, to = F_max, length.out = no_steps)}.}

\item{F_max}{The maximum fishing mortality. Used only if \code{F_range} is missing.}

\item{F_min}{The minimum fishing mortality. Used only if \code{F_range} is missing.}

\item{no_steps}{The number of steps to use. Only used if \code{F_range} is missing.}

\item{distance_func}{A function that will be called after every \code{t_per} years
with both the previous and the new state and that should return a number
that in some sense measures the distance between the states. By default
this uses the function \code{\link[=distanceSSLogN]{distanceSSLogN()}} that you can use as a model for your
own distance function.}

\item{tol}{The \code{projectToSteady} function stops when the relative change in the egg production
RDI over t_per years is less than tol for every species.}

\item{t_max}{The longest time to run project to find steady state.}
}
\value{
A data frame with columns \code{F} and \code{yield}.
}
\description{
Calculates the yield of a species for a range of fishing mortalities for
that species while the fishing mortalities for the other species are held
fixed.
}
\examples{
\dontrun{
params <- newMultispeciesParams(NS_species_params_gears, inter)
y <- getYieldVsF(params, "Cod", F_max = 1, no_steps = 5)
}
}
\seealso{
plotYieldVsF
}
\concept{summary functions}
\concept{summary_function}
