% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthTab.R
\name{growthTab}
\alias{growthTab}
\alias{growthTabUI}
\title{Growth tab for tuning gadget
inheritParams biomassTab}
\usage{
growthTab(input, output, session, params, logs, ...)

growthTabUI(...)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding a MizerParams object.}

\item{logs}{Environment holding the log of steady states.}

\item{...}{Unused}
}
\description{
Growth tab for tuning gadget
inheritParams biomassTab
}
