% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionControl.R
\name{interactionControl}
\alias{interactionControl}
\alias{interactionControlUI}
\title{Controlling the interaction matrix in the tuning gadget}
\usage{
interactionControl(input, output, session, params, flags, ...)

interactionControlUI(p, sp)
}
\arguments{
\item{input}{.}

\item{output}{.}

\item{session}{.}

\item{params}{.}

\item{flags}{.}

\item{...}{.}

\item{p}{The MizerParams object currently being tuned.}

\item{sp}{The single row from the pecies parameter data frame for the
currently selected species.}
}
\description{
Controlling the interaction matrix in the tuning gadget
}
