% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchGrowth.R
\name{matchGrowth}
\alias{matchGrowth}
\title{Adjust model to produce observed growth}
\usage{
matchGrowth(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
A modified MizerParams object with rescaled search volume, maximum
consumption rate and metabolic rate and rescaled species parameters
\code{gamma},\code{h}, \code{ks} and \code{k}.
}
\description{
Scales the search volume, the maximum consumption rate and the metabolic rate
all by the same factor in order to achieve a growth rate that allows
indivdiuals to reach their maturity size by their maturity age while keeping
the feeding level and the critical feeding level unchanged.
}
\details{
Maturity size and age are taken from the \code{w_mat} and \code{age_mat} columns in the
species_params data frame.
}
