% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchYield.R
\name{matchYield}
\alias{matchYield}
\title{Match observed yields}
\usage{
matchYield(params, keep = c("egg", "biomass", "number"))
}
\arguments{
\item{params}{A MizerParams object}

\item{keep}{A string determining which quantity is to be kept constant. The
choices are "egg" which keeps the egg density constant, "biomass" which
keeps the total biomass of the species constant and "number" which keeps
the total number of individuals constant.}
}
\value{
A MizerParams object with updated catchabilities
}
\description{
This function matches the observed yields of all the gears for all the
species by scaling the catchabilities by the ratio between current modelled
yield and observed yield.
}
