% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchYield.R
\name{matchYield}
\alias{matchYield}
\title{Match observed yields}
\usage{
matchYield(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
A MizerParams object with updated catchabilities
}
\description{
This function matches the observed yields of all the gears for all the
species by scaling the catchabilities by the ratio between current modelled
yield and observed yield.
}
