% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherControl.R
\name{otherControl}
\alias{otherControl}
\alias{otherControlUI}
\title{Controlling the species parameters not included in other controls}
\usage{
otherControl(input, output, session, params, flags, ...)

otherControlUI(p, sp)
}
\arguments{
\item{input}{.}

\item{output}{.}

\item{session}{.}

\item{params}{.}

\item{flags}{.}

\item{...}{.}

\item{p}{The MizerParams object currently being tuned.}

\item{sp}{The single row from the pecies parameter data frame for the
currently selected species.}
}
\description{
Controlling the species parameters not included in other controls
}
