% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherControl.R
\name{otherControl}
\alias{otherControl}
\alias{otherControlUI}
\title{Controlling the species parameters not included in other controls}
\usage{
otherControl(input, output, session, params, params_old, flags, ...)

otherControlUI(p, input)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{params_old}{Reactive value holding non-updated MizerParams object}

\item{flags}{Environment holding flags to skip certain observers}

\item{...}{Unused}

\item{p}{The MizerParams object currently being tuned.}
}
\description{
Adds controls for the parameters that did not fit into other categories.
}
\details{
The parameters included are:
\itemize{
\item conversion efficiency \code{alpha}
\item metabolic rate coefficient \code{ks} and exponent \code{p}
\item activity rate coefficient \code{k}
\item external mortality at maturity size
}

If the external mortality at maturity size is changed, then the
external mortality at all other sizes is scaled by the same factor.

If the metabolism exponent \code{p} is changed, this also changes \code{ks} so
that metabolic rate at maturity stays the same.
}
