% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBiomassFlux.R
\name{plotBiomassFlux}
\alias{plotBiomassFlux}
\title{Plot biomass flux}
\usage{
plotBiomassFlux(params, ylim = c(NA, NA), total = FALSE)
}
\arguments{
\item{params}{A MizerParams object}

\item{ylim}{A numeric vector of length 2 giving the lower and upper limits
for the y-axis. If the lower limit is NA, it is set to 1e-20. If the upper
limit is NA, it is set to the maximum value of the biomass flux.
Default is c(NA, NA).}

\item{total}{A logical indicating whether to add a line for the total biomass
flux. Default is FALSE.}
}
\description{
Plot biomass flux
}
\examples{
\dontrun{
  params <- newMultispeciesParams(NS_species_params_gears, inter)
  plotBiomassFlux(params)
}
}
