% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpectraRelative.R
\name{plotBiomassRelative}
\alias{plotBiomassRelative}
\alias{plotlyBiomassRelative}
\title{Plot change in biomass over time}
\usage{
plotBiomassRelative(sim, ...)

plotlyBiomassRelative(sim, ...)
}
\arguments{
\item{sim}{A MizerSim object}

\item{...}{Parameters passed to \code{getBiomass()}}
}
\description{
Plot change in biomass over time
}
