% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRelative.R
\name{plotBiomassRelative}
\alias{plotBiomassRelative}
\alias{plotlyBiomassRelative}
\title{Plot change in biomass over time}
\usage{
plotBiomassRelative(sim, sim_original = NULL, ...)

plotlyBiomassRelative(sim, sim_original = NULL, ...)
}
\arguments{
\item{sim}{A MizerSim object}

\item{sim_original}{Another MizerSim object to compare the biomasses to. If
NULL (default) then the biomasses are compared to the initial biomasses in
\code{sim}.}

\item{...}{Parameters passed to \code{getBiomass()}}
}
\description{
Plot change in biomass over time
}
\examples{
plotBiomassRelative(NS_sim)
}
