% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRelative.R
\name{plotBiomassRelative}
\alias{plotBiomassRelative}
\alias{plotlyBiomassRelative}
\title{Plot change in biomass over time}
\usage{
plotBiomassRelative(sim, sim_original = NULL, species = NULL, ...)

plotlyBiomassRelative(sim, sim_original = NULL, species = NULL, ...)
}
\arguments{
\item{sim}{A MizerSim object}

\item{sim_original}{Another MizerSim object to compare the biomasses to. If
NULL (default) then the biomasses are compared to the initial biomasses in
\code{sim}.}

\item{species}{The species to be selected. Optional. By default all target
species are selected. A vector of species names, or a numeric vector with
the species indices, or a logical vector indicating for each species
whether it is to be selected (TRUE) or not.}

\item{...}{Parameters passed to \code{getBiomass()}}
}
\description{
Plot change in biomass over time
}
\examples{
plotBiomassRelative(NS_sim)
plotBiomassRelative(NS_sim, species = c("Cod", "Sole"))
}
