% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBiomassVsSpecies.R
\name{plotBiomassVsSpecies}
\alias{plotBiomassVsSpecies}
\alias{plotBiomassVsSpecies.MizerParams}
\title{Plot the biomass against species}
\usage{
plotBiomassVsSpecies(params, ...)

\method{plotBiomassVsSpecies}{MizerParams}(params, ...)
}
\arguments{
\item{params}{A MizerParams object}

\item{...}{Not used.}
}
\description{
Plot the biomass against species
}
