% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBiomassVsSpecies.R
\name{plotBiomassVsSpecies}
\alias{plotBiomassVsSpecies}
\title{Plot the biomass against species}
\usage{
plotBiomassVsSpecies(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\description{
Plot the biomass against species
}
