% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNumberVsSpecies.R
\name{plotNumberVsSpecies}
\alias{plotNumberVsSpecies}
\title{Plot the number against species}
\usage{
plotNumberVsSpecies(params, ...)
}
\arguments{
\item{params}{A MizerParams object}

\item{...}{Not used.}
}
\description{
This is a generic function with a method for objects of class
\link[mizer:MizerParams]{MizerParams}.
}
\examples{
plotNumberVsSpecies(NS_params)

}
