% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNumberVsSpecies.R
\name{plotNumberVsSpecies}
\alias{plotNumberVsSpecies}
\title{Plot the number against species}
\usage{
plotNumberVsSpecies(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\description{
Plot the number against species
}
