% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotResourceLevel}
\alias{plotResourceLevel}
\title{Plot the proportion of the resource abundance compared to
the resource carrying capacity}
\usage{
plotResourceLevel(object, return_data = FALSE, ...)
}
\arguments{
\item{object}{An object of class \linkS4class{MizerSim} or
\linkS4class{MizerParams}.}

\item{return_data}{A boolean value that determines whether the formatted data
used for the plot is returned instead of the plot itself. Default value is
FALSE}

\item{...}{Other arguments (currently unused)}
}
\value{
A ggplot2 object, unless \code{return_data = TRUE}, in which case a data
frame with the three variables 'w', 'value', 'Resource' is returned.
}
\description{
This is a generic function with methods for objects of class
\linkS4class{MizerSim} and \linkS4class{MizerParams}. The method for a
\code{MizerSim} object plots the resource level in the initial state of the
simulation.
}
\examples{
\donttest{
plotResourceLevel(NS_params)

# Returning the data frame
fr <- plotResourceLevel(NS_params, return_data = TRUE)
str(fr)
}
}
\seealso{
\link[mizer:plotting_functions]{plotting_functions}

Other plotting functions:
\code{\link[=plotDeath]{plotDeath()}},
\code{\link[=plotEnergyBudget]{plotEnergyBudget()}},
\code{\link[=plotResourcePred]{plotResourcePred()}},
\code{\link[=plotYieldVsF]{plotYieldVsF()}},
\code{\link[=plotYieldVsSize]{plotYieldVsSize()}}
}
\concept{plotting functions}
