% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotResourceLevel}
\alias{plotResourceLevel}
\title{Plot the proportion of the resource spectrum(s) compared to
their carrying capacity}
\usage{
plotResourceLevel(object, return_data = FALSE)
}
\arguments{
\item{object}{An object of class \linkS4class{MizerSim} or
\linkS4class{MizerParams}.}

\item{return_data}{A boolean value that determines whether the formatted data
used for the plot is returned instead of the plot itself. Default value is
FALSE}
}
\value{
A ggplot2 object, unless \code{return_data = TRUE}, in which case a data
frame with the three variables 'w', 'value', 'Resource' is returned.
}
\description{
Plot the proportion of the resource spectrum(s) compared to
their carrying capacity
}
\examples{
\donttest{
plotResourceLevel(NS_params)

# Returning the data frame
fr <- plotResourceLevel(NS_params, return_data = TRUE)
str(fr)
}
}
\seealso{
\link{plotting_functions}

Other plotting functions: 
\code{\link{plotDeath}()},
\code{\link{plotEnergyBudget}()},
\code{\link{plotResourcePred}()},
\code{\link{plotYieldCurve}()},
\code{\link{plotYieldVsF}()},
\code{\link{plotYieldVsSize}()}
}
\concept{plotting functions}
