% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpectra2.R
\name{plotSpectra2}
\alias{plotSpectra2}
\title{Show two size spectra in the same plot}
\usage{
plotSpectra2(
  object1,
  object2,
  name1 = "First",
  name2 = "Second",
  power = 1,
  ...
)
}
\arguments{
\item{object1}{First MizerParams or MizerSim object.}

\item{object2}{Second MizerParams or MizerSim object.}

\item{name1}{An optional string with the name for the first model, to be used
in the legend. Set to "First" by default.}

\item{name2}{An optional string with the name for the second model, to be
used in the legend. Set to "Second" by default.}

\item{power}{The abundance is plotted as the number density times the weight
raised to this power. The default power = 1 gives the biomass density,
whereas power = 2 gives the biomass density with respect to logarithmic
size bins.}

\item{...}{Parameters to pass to \code{plotSpectra()}}
}
\value{
A ggplot2 object
}
\description{
Show two size spectra in the same plot
}
\examples{
sim1 <- project(NS_params, t_max = 10)
sim2 <- project(NS_params, effort = 0.5, t_max = 10)
plotSpectra2(sim1, sim2, "Original", "Effort = 0.5")
}
