% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpectraRelative.R
\name{plotSpectraRelative}
\alias{plotSpectraRelative}
\title{Plot the relative difference between two spectra}
\usage{
plotSpectraRelative(object1, object2, ...)
}
\arguments{
\item{object1}{An object of class MizerSim or MizerParams}

\item{object2}{An object of class MizerSim or MizerParams}

\item{...}{Parameters passed to \code{plotSpectra()}}
}
\value{
A ggplot2 object
}
\description{
If we denote the number density from the first object as \eqn{N_1(w)} and
that from the second object as \eqn{N_2(w)}, then this plot shows
\deqn{(N_2(w) - N_1(w)) / (N_2(w) - N_1(w)).}
}
\details{
Note that it does not matter whether the relative difference is calculated
for the number density or the biomass density or the biomass density in log
weight because the factors of \eqn{w} by which the densities differ cancels
out in the relative difference.
}
\examples{
sim1 <- project(NS_params, t_max = 10)
sim2 <- project(NS_params, effort = 0.5, t_max = 10)
plotSpectraRelative(sim1, sim2)
}
