% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpectraRelative.R
\name{plotYieldRelative}
\alias{plotYieldRelative}
\alias{plotlyYieldRelative}
\title{Plot change in yield over time}
\usage{
plotYieldRelative(sim, params, ...)

plotlyYieldRelative(sim, params, ...)
}
\arguments{
\item{sim}{A MizerSim object}

\item{params}{A MizerParams object describing the initial state, used to
calculate the initial yield.}

\item{...}{Parameters passed to \code{getYield()}}
}
\description{
Plot change in yield over time
}
