% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpectraRelative.R
\name{plotYieldRelative}
\alias{plotYieldRelative}
\alias{plotlyYieldRelative}
\title{Plot change in yield over time}
\usage{
plotYieldRelative(sim, object_original, ...)

plotlyYieldRelative(sim, object_original, ...)
}
\arguments{
\item{sim}{A MizerSim object}

\item{object_original}{A MizerParams or MizerSim object to calculate
differences from.}

\item{...}{Parameters passed to \code{getYield()}}
}
\description{
Plot change in yield over time
}
