% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYieldVsFcurve.R
\name{plotYieldVsF}
\alias{plotYieldVsF}
\title{Plot a yield-vs-F curve}
\usage{
plotYieldVsF(
  params,
  species,
  no_steps = 10,
  F_max = 3,
  F_range = seq(0, F_max, length.out = no_steps)
)
}
\arguments{
\item{params}{An object of class \code{MizerParams}.}

\item{species}{Name of the target species}

\item{no_steps}{The number of steps to use. Only used if \code{F_range} is missing.}

\item{F_max}{The maximum fishing mortality. Used only if \code{F_range} is missing.}

\item{F_range}{A sequence of fishing mortalities at which to evaluate the
yield. If missing, it is set to
\code{seq(from = 0, to = F_max, length.out = no_steps)}.}
}
\value{
A ggplot object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculates the yield of a species for a range of fishing mortalities for
that species while the fishing mortalities for the other species are held
fixed.
}
\examples{
\dontrun{
params <- newMultispeciesParams(NS_species_params_gears, inter)
plotYieldVsF(params, "Cod")
}
}
\seealso{
getYieldVsF

Other plotting functions: 
\code{\link{animateSpectra}()},
\code{\link{displayFrames}()}
}
\concept{plotting functions}
