% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYieldVsFcurve.R
\name{plotYieldVsF}
\alias{plotYieldVsF}
\title{Plot a yield-vs-F curve}
\usage{
plotYieldVsF(
  params,
  species,
  no_steps = 10,
  F_max,
  F_range,
  effort_it = 1,
  distance_func = distanceSSLogN,
  tol = 0.001,
  max_func = maxFthreshold,
  threshold_var = 10,
  t_max = 100
)
}
\arguments{
\item{params}{An object of class \code{MizerParams}.}

\item{species}{Name of the target species}

\item{no_steps}{The number of steps to use. Only used if \code{F_range} is missing.}

\item{F_max}{The maximum fishing mortality. Used only if \code{F_range} is missing.}

\item{F_range}{A sequence of fishing mortalities at which to evaluate the
yield. If missing, it is set to
\code{seq(from = 0, to = F_max, length.out = no_steps)}.}

\item{effort_it}{The width of steps used to find maximum fisheries mortality,
higher values means a faster calculation but less accurate maximum.}

\item{distance_func}{A function that will be called after every \code{t_per} years
with both the previous and the new state and that should return a number
that in some sense measures the distance between the states. By default
this uses the function \code{\link[=distanceSSLogN]{distanceSSLogN()}} that you can use as a model for your
own distance function.}

\item{tol}{The \code{projectToSteady} function stops when the relative change in the egg production
RDI over t_per years is less than tol for every species.}

\item{max_func}{A function used in \code{getMaxF} to determines the threshold to use
when stopping the fisheries effort.}

\item{threshold_var}{A numeric value used in max_func to calculate the threshold
from the \code{params} object. If supplied, the function returns the value of the
threshold using \code{threshold_var}. If not supplied, the function returns the
unchanged value, which can be compared against the version which uses \code{threshold_var}.}

\item{t_max}{The longest time to run project to find steady state.}
}
\value{
A ggplot object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculates the yield of a species for a range of fishing mortalities for
that species while the fishing mortalities for the other species are held
fixed.
}
\examples{
\dontrun{
params <- newMultispeciesParams(NS_species_params_gears, inter)
plotYieldVsF(params, "Cod")
}
}
\seealso{
getYieldVsF

Other plotting functions: 
\code{\link{plotDeath}()},
\code{\link{plotEnergyBudget}()},
\code{\link{plotResourceLevel}()},
\code{\link{plotResourcePred}()},
\code{\link{plotYieldVsSize}()}
}
\concept{plotting functions}
