% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotYieldVsSize}
\alias{plotYieldVsSize}
\alias{plotlyYieldVsSize}
\title{Plot the size distribution of the catch}
\usage{
plotYieldVsSize(
  object,
  species = NULL,
  catch = NULL,
  x_var = c("Weight", "Length"),
  return_data = FALSE
)

plotlyYieldVsSize(
  object,
  species = NULL,
  catch = NULL,
  x_var = c("Weight", "Length"),
  ...
)
}
\arguments{
\item{object}{An object of class \linkS4class{MizerSim} or
\linkS4class{MizerParams}.}

\item{species}{The name of the predator species for which to plot the
mortality.}

\item{catch}{Optional. Data frame holding binned observed catch data. The data can
be binned either into length bins or weight bins. In the former case the data
frame should have columns \code{length} and \code{dl} holding the start of
the size bins in cm and the width of the size bins in cm respectively. In
the latter case the data frame should have columns \code{weight} and
\code{dw} holding the start of the size bins in grams and the width of the
size bins in grams. The data frame also needs to have the columns
\code{species} (the name of the species), \code{catch} (the number of
individuals of a particular species caught in a size bin).}

\item{x_var}{Determines whether to show the size distribution of the catch as
a function of weight ("Weight") or as a function of length ("Length").
Default is "Weight".}

\item{return_data}{A boolean value that determines whether the formatted data
used for the plot is returned instead of the plot itself. Default value is
FALSE}

\item{...}{Other arguments (currently unused)}
}
\value{
A ggplot2 object, unless \code{return_data = TRUE}, in which case a list
composed of two slots is returned. First slot is a data frame with the four
variables 'w' or 'l' (depending on \code{x_var}), 'Catch density', 'Type', 'Species
and the second slot is a data frame with the four variables 'w_mat',
'Species', 'y_coord', 'Type' (to plot vertical lines).
}
\description{
Plots the normalised number density of the catch for a species as a function
of either length or weight. In addition to the catch in the model, also the
observed catch will be plotted if it is supplied via the \code{catch} argument.
Also superimposes a plot of the number density of all individuals of the
species.
}
\examples{
\donttest{
plotYieldVsSize(NS_params, species = "Cod")

# Returning the data frame
fr <- plotYieldVsSize(NS_params, species = "Cod", return_data = TRUE)
str(fr)
}
}
\seealso{
\link{plotting_functions}

Other plotting functions: 
\code{\link{plotDeath}()},
\code{\link{plotEnergyBudget}()},
\code{\link{plotResourcePred}()},
\code{\link{plotResource}()},
\code{\link{plotYieldVsF}()}
}
\concept{plotting functions}
