% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotYieldVsSpecies.R
\name{plotYieldVsSpecies}
\alias{plotYieldVsSpecies}
\title{Plot the yield against species}
\usage{
plotYieldVsSpecies(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\description{
Plot the yield against species
}
