% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predationControl.R
\name{predationControl}
\alias{predationControl}
\alias{predationControlUI}
\title{Controlling the predation parameters in the tuning gadget}
\usage{
predationControl(input, output, session, params, params_old, flags, ...)

predationControlUI(p, input)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{params_old}{Reactive value holding non-updated MizerParams object}

\item{flags}{Environment holding flags to skip certain observers}

\item{...}{Unused}

\item{p}{The MizerParams object currently being tuned.}
}
\description{
Controlling the predation parameters in the tuning gadget
}
