% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulateParams.R
\name{removeBackgroundSpecies}
\alias{removeBackgroundSpecies}
\title{Remove all background species}
\usage{
removeBackgroundSpecies(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\description{
Remove all background species
}
