% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproductionControl.R
\name{reproductionControl}
\alias{reproductionControl}
\alias{reproductionControlUI}
\title{Controlling the reproduction parameters in the tuning gadget}
\usage{
reproductionControl(input, output, session, params, flags, ...)

reproductionControlUI(p, sp)
}
\arguments{
\item{input}{.}

\item{output}{.}

\item{session}{.}

\item{params}{.}

\item{flags}{.}

\item{...}{.}

\item{p}{The MizerParams object currently being tuned.}

\item{sp}{The single row from the pecies parameter data frame for the
currently selected species.}
}
\description{
Controlling the reproduction parameters in the tuning gadget
}
