% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulateParams.R
\name{scaleDownBackground}
\alias{scaleDownBackground}
\title{Scale background down by a factor}
\usage{
scaleDownBackground(params, factor)
}
\arguments{
\item{params}{A MizerParams object}

\item{factor}{A number giving the factor by which the background abundance
will be reduced}
}
\description{
Scale background down by a factor
}
