% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneParams.R
\name{tuneGrowth}
\alias{tuneGrowth}
\title{Tune Growth}
\usage{
tuneGrowth(params, match = c("biomass", "yield", "none"))
}
\arguments{
\item{params}{MizerParams object to tune. If missing, the gadget tries to recover
information from log files left over from aborted previous runs.}

\item{match}{Determines whether the biomasses or the yields should be
matched to observations each time the "steady" button is pressed.}
}
\value{
The tuned MizerParams object
}
\description{
A simplified instance of \code{tuneParams()} that is useful for tuning growth
rates.
}
