% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYieldVsFcurve.R
\name{yieldCalculator}
\alias{yieldCalculator}
\title{Calculates yield from steady sim}
\usage{
yieldCalculator(params, effort_vec, idx_species)
}
\arguments{
\item{params}{Mizer params object}

\item{effort_vec}{vector of effort values to calcualte the yield from}

\item{idx_species}{targeted species row number in \code{params@species_params}}
}
\value{
a vector of yield value of same length as \code{effort_vec}
}
\description{
This function replaces a loop used multiple times within
\code{getYieldVsCurve}
}
